[feature/mainpage] feat : 네이버 지도 GeoCoding 파트 (완료 전)

+++ lib/main_page/controllers/shortWeatherController.dart
@@ -73,7 +73,7 @@ class ShortWeatherController extends GetxController {
   }
 
   //자외선 지수 세기
-  void indexUvi(double uvi){
+  void indexUvi(int uvi){
     if(uvi < 3){
       crnt_uvi.value = "낮음";
     }
@@ -101,7 +101,7 @@ class ShortWeatherController extends GetxController {
         city.value = "현재 위치"; // 추후 추가 예정
         tempMax.value = "${weatherData["daily"][0]["temp"]["max"].round()}°";
         tempMin.value = "${weatherData["daily"][0]["temp"]["min"].round()}°";
-        double uvi = weatherData["current"]["uvi"];
+        int uvi = weatherData["current"]["uvi"];
         indexUvi(uvi);
 
         weeklyWeather.value = List.generate(7, (index) {
